+++ Editor/MenuActions/Object/MergeObjects.cs
@@ -6,6 +6,7 @@ using UnityEngine;
 using UnityEditor;
 using UnityEditor.ProBuilder.UI;
 using UnityEngine.ProBuilder.MeshOperations;
+using UnityEngine.ProBuilder.Shapes;
 
 namespace UnityEditor.ProBuilder.Actions
 {
@@ -42,6 +43,12 @@ namespace UnityEditor.ProBuilder.Actions
             if (MeshSelection.selectedObjectCount < 2)
                 return new ActionResult(ActionResult.Status.Canceled, "Must Select 2+ Objects");
 
+            DoMergeObjectsAction();
+            return new ActionResult(ActionResult.Status.Success, "Merged Objects");
+        }
+
+        internal List<ProBuilderMesh> DoMergeObjectsAction()
+        {
             var selected = MeshSelection.top.ToArray();
             ProBuilderMesh currentMesh = MeshSelection.activeMesh;
             UndoUtility.RecordObject(currentMesh, "Merge Objects");
@@ -56,8 +63,18 @@ namespace UnityEditor.ProBuilder.Actions
                     {
                         mesh.gameObject.name = Selection.activeGameObject.name + "-Merged";
                         UndoUtility.RegisterCreatedObjectUndo(mesh.gameObject, "Merge Objects");
+                        
                         Selection.objects = res.Select(x => x.gameObject).ToArray();
                     }
+                    
+                    // Remove PolyShape and ProBuilderShape components if any are present post-merge
+                    var polyShapeComp = mesh.gameObject.GetComponent<PolyShape>();
+                    if (polyShapeComp != null )
+                        UndoUtility.DestroyImmediate(polyShapeComp);
+                    
+                    var proBuilderShape = mesh.gameObject.GetComponent<ProBuilderShape>();
+                    if (proBuilderShape != null )
+                        UndoUtility.DestroyImmediate(proBuilderShape);
                 }
 
                 // Delete donor objects if they are not part of the result
@@ -68,9 +85,8 @@ namespace UnityEditor.ProBuilder.Actions
                 }
             }
 
-            ProBuilderEditor.Refresh();
+            return res;
 
-            return new ActionResult(ActionResult.Status.Success, "Merged Objects");
         }
     }
 }